fix sorting by average cam range
Storing cam ranges as doubles in db was causing them to be serialized as
strings. Floats are serialized as floats. Cam ranges only need float
precision so the simplest solution was to change the db data type.

+++ db/migrate/20200205133047_create_cams.rb
@@ -5,10 +5,10 @@ class CreateCams < ActiveRecord::Migration[6.0]
     create_table :cams do |t|
       t.string :name, null: false
       t.references :cam_style, foreign_key: true, null: false
-      t.decimal :range_min, null: false
-      t.decimal :range_max, null: false
-      t.decimal :weight, null: false
-      t.decimal :strength, null: false
+      t.float :range_min, null: false
+      t.float :range_max, null: false
+      t.float :weight, null: false
+      t.float :strength, null: false
       t.string :color, null: false
     end
   end
